+++ CHANGELOG.md
@@ -10,6 +10,7 @@ CHANGELOG for FlatCAM beta
 20.03.2022
 
 - added a last resort option to load old projects; the result is not guaranteed if the differences are too great
+- updated the code in FlatCAMCNCJob object such that more attributes will be serialized in order for loading a project correctly when it has CNCJob objects
 
 19.03.2022